refactor: add coloring to help
