+++ src/streams/inputStream.js
