+++ src/streams/outputStream.js
